Fix inability to detect already bumped (#1)
Something in Git has changed, such that tag messages generated tag refs
different than the commit they were tagging. This caused the script to
bump even if a tag already existed for HEAD.

Removal of the tag message and annotation reverted the behavior to how
git worked when v0.1.0 was released.